DLeyanDemo: get started on GVariant, GVariantType
This is the initial implementation of the Java wrapper classes for
GVariant and GVariantType classes, and the native code supporting it.
It only supports the GVariant types required for the renderer service.
The java code is complete. The native code is not quite complete;
functions concerning arrays are not implemented. Nothing has been
tested.

Now using javah to generate .h files declaring native methods found
in .java files. This should be done automatically, but for now it's
done by hand, because I don't know how to put it into the build.