DLeynaDemo: pointer <=> jlong conversion macros
This defines and uses a couple of macros for converting between
pointers and jlong in native code. If you just cast from pointer
to jlong (or vice-versa) you get a compile-time warning in the case
where the java VM uses a 32-bit address space. The macros avoid the
warning by doing an intermediate cast to size_t.