NativeLibs: Quick fix for gunpn/-av symbol clashes
This allows gupnp and gupnp-av to be bundled on a single
shared library.

The situation is quite broken upstream. Real solution would
likely be to make sure both implementations are on-par
functionality wise and make gupnp-av use the routines from
gupnp. Maybe an independent library solution can be explored
too but none of these fixes can be implemented quick enough.
Besides, this has to be discussed with upstream so it's
unlikely that we will be in position to implement a proper
and upstreamable fix during this project's life time.