DLeynaDemo: Garnish the Renderer .aidl files.
Fill in IRendererService.aidl and IRendererCallback.aidl interface
definitions. Move them from the service package to the library
package, because both the services and the apps will need them.
Fill in the corresponding stubs that they translate to with skeleton
code.

Implement the Parcelable interface for the Icon type so that can be
used with Binder.

Switch to using Bundle as the type for the metadata in the API method
IRendererControllerEvents.onMetadataChanged(). A Bundle is a map from
strings to objects that can be of a variety of types, which is just
what we need, and it is all set up to work with Binder.

Also a few corrections to the documentation.