DLeynaDemo: Implement library-level callbacks.
This implements all the event callbacks at the library level: all
RendererManager.Events and IRendererControllerEvents callbacks. The
callbacks all originate in binder threads, and are transfered to the
app's main thread by using a Handler.

Also added a clientUnregister() to the RendererService.

Also minor fixes to code and documentation, one case of moving a block
of code to improve readability.