DLeynaDemo: Add push activity.
This adds an Activity to handle two-box push initiated by the user
from another application. For example, the user selects a photo or
video in the gallery app and does a "share" operation.

This new Activity is just a blank screen for now, but it does get
launched correctly in the above scenario and it does connect to the
background renderer service.