DLeynaDemo: Get connector barely working.
This gets the native connector working to the point where we are able
to call dleyan_main_loop_start() and that successfully loads the
connector plugin and calls the plugin's initialize().

- Build the native connector plugin as as a separate shared library
because dleyna wants to load it at runtime. But in fact we have to
load it ourselves at runtime before dleyna, otherwise loading the
big shared library fails.

- In RendererService.onBind() create a daemon thread and call
dleyna_main_loop_start() on that thread. This is the thread that
all the dleyna stuff runs on.

- Clarify in comments that the g_main_loop runs on a thread that is NOT
the main thread in the Android sense.

- Rename interface Connector => IConnector. Rename a couple of .c files.

- Note in a comment that we will have to move the call JNI.cleanTempDir()
some day.