DLeynaDemo: More connector work.
This gets the connector stuff working to the point where it handles
the first publish_object call informing us of the existence of the
renderer manager object. Traces indicate that gssdp is finding network
interfaces.

Also fixes a threading problem that would show up if you exited the
app and restarted it almost immediately. The main thread in the
android service now waits (in the onUnbind()) until the daemon thread
(that runs the g_main_loop) has exited.