DLeynaDemo: Connector needs config file.
When the native dleyna-renderer service starts up, it tries to read
configuration info from ~/.conf/dleyna-renderer-service.conf. Among
the config parameters is the type of connector, 'dbus' or 'android',
and since the default is 'dbus' we really need the config file on the
device. (Until now we had just installed it by hand onto the device.)

This change adds that file as an asset to the app package, and copies
it to the right place at jni initialization time if it's not already
there.

Also: remove unused .h file.