DLeynaDemo: Improve connect/disconnect robustness.
This adds 'connect' and 'disconnect' buttons to the UI, and does the
RenderManager API connect/disconnect upon their click events, instead
of from onCreate() and onDestroy() as an app usually would. All this
so we can validate the connect/disconnect code. Surpises awaited us.
For example, unbinding the only client of the service does *not*
cause an onServiceDisconnected() callback. Also, if you kill the
service process the sytem will relaunch it after a few seconds, and
apparently re-bind the activity to it, but I haven't figured out how
to handle that correctly yet. Anyway, things are a bit more robust
now.