DLeynaDemo: Implement getRenderers().
This implements RendererManager.getRenderers() from the API all
the way down through all the turtles and back up again. Highlights:

* Add a UI button for exercising RendererManager.getRenderers().

* All the gory details in RendererService for transfering the
call from a binder thread to the g_main_loop, waiting for the
asynchronous response from the native level, and propagating
the response back up.

* Add a new class, GMainLoop, and its native peer, for queueing
Runnable objects to execute on the g_main_loop. Use an instance of
this class in Connector.

* Add a method to GVariant to extract an element at a given index
from a container. (The parameters to, and the response from, the
native renderer service are wrapped in a GVariant tuple.)

* Since the implementation will be the same for both DMR and DMS
services, put the implementation of IConnector in a separate class,
Connector, and use an instance of that in RendererService instead of
implementing the interface there. Remove IConnector.

* Rename interface IRendererCallback => IRendererClient, and add an
IRendererClient arg to every method in the IRendererService interface.