DLeynaDemo: First connector notifications working.
This gets a first couple of connector notifications, RendererFound and
RendererLost, working all the way up to the application level.

Added an interface, IConnectorClient, and an implementation of it in
RendererService, so as to keep the Connector object independent of the
type of service.  Connector gets an instance of IConnectorClient in
its constructor, and uses it to pass notifications up to the service.
The service decodes the notification, and decides whether and how to
send it up.

Added some more GVariant stuff: extract a child from a container.