DLeynaDemo: IController events working at all levels.
This gets all of the Renderer IController events (aka "property change
notifications") working, from the connnector all the way up to the
application. One exception: the Metadata event isn't implemented
because I don't yet understand it, and the doc isn't much help.

Controller property change notifications come out of the connector in
batches. We had originally started to implement this using a scheme
where each individual property change was sent to the client by binder
IPC. But now we send the whole batch, wrapped in a Bundle, in one IPC
transaction.