DLeynaDemo: Implement IRendererDevice methods.
This gets all of the IRendererDevice methods working, except for
getIcon() and cancel().

Also: fix a bug where we were passing the *java* version of the
GVariant params to the native dispatch function, instead of the
*native* version.

Also: fix GVariant.getArrayOfDouble(). It's *double* not *long*.

Also: make all the methods of the RemoteObject.Store synchronized
since it's accessed from multiple threads.