DLeynaDemo: Implement IRendererDevice.cancel().
This implements the method that cancels any pending method invocations
on a given renderer.

The 'List' button now runs a background thread that sequentially calls
several methods on all renderers, and forwards the results of each
call to the main thread to be displayed. Meanwhile, if you're fast
enough, you can press the 'Cancel' button to send a cancel() to the
renderer service from the main thread.

I verified by examining traces that if the cancel request gets to the
connector while there is another invocation pending (not always the
case, depending on when you happend to press the button, etc) that
pending request is indeed cancelled.

Also: fixes an error in the initial implementation of the native peer
of the GMainLoop class. The gMainLoop.idleAdd() could only work if
there was never more than one pending invocation to run, but now it
works right. Doh!