DLeynaDemo: Add proper error reporting.
This change adds proper error reporting of failed method invocations,
from the connector up to the API and the app. We verified that
cancelled pending operations at the native service level are correctly
reported by the app.

At the API level, define a new exception, DLeynaException, and declare
that all relevant API methods throw it. In the app, catch it, and
write the error message and stack trace to the log.

Update the API doc appropriately.

Implement the native code for getting stuff from a GError. Use it in
Connector to pass the error code and message up to the service. In the
service, pass it up to the client library in a Bundle. In the library,
if the error info is in the Bundle, throw the exception.

Also: remove a false "throws RemoteException".