DLeynaDemo: Fix the manager object bug.
This fixes a problem that showed up when using a more recent version
of the native dleyna-renderer. We had presumed the connector would
only announce one interface with isRoot == true, but the manager
object can have multiple interfaces and they are all announced with
isRoot==true.

So now we don't presume that, and we don't use the isRoot flag at all,
and we don't cache the manager object at all, we just look it up by
object path name and interface name.