DLeynaDemo: Prettify the push activity.
This makes the PushActivity look nicer -- you still see the activity
that launched it, but darkened, beneath a transparent background. The
spinning wheel has words beneath it. The discovered devices are
vertically centered.

Also, handle device orientation change, which means that when the
user rotates the device, the activity is completely relaunched for
the new orientation. And that means that we have to reconnect to the
service and restart the discovery procedure. Currently, if you had
already discovered something before rotating the device, you will
see the spinning wheel again briefly. It would be possible, but not
entirely trivial, to remember the previously discovered devices so as
to avoid the dreaded wheel. But what if they really disappeared from
the network between the two incarnations of the activity?

Also, add a means of faking discovered renderers, for testing the UI.