DLeynaDemo: Split PushActivity into two activites.
In preparation for adding a controller UI, split PushActivity into two
parts. The PushActivity now launches the new RendererChoiceActivity to
to do the discovery, and gets back the renderer chosen by the user. It
still displays/plays the media the dumb way -- play until the user
navigates away (or rotates the device...).

Splitting has a nice side effect: rotating the device during discovery
after having found renderers no longer results in the spinning wheel
being redisplayed for a few seconds. Yay! This is because each
activity has a connection to the background service, and when you
rotate the device, they are shut down and relaunched in an order that
results in at least one of them always having a connection, so the
service is never shut down.