DLeynaDemo: PushActivity: A first controller UI.
This is the first shot at a controller UI for the push activity: a
progress bar that represents the position of the playback, and a
play/pause button. When playback is paused, the user can slide the
knob on the progress bar to set a new position for playback.

The animation of the progress bar is pretty jerky, especially for
short video clips, because we're polling the renderer once per
second to get the position, and not trying to do any fancy smoothing.