DLeynaDemo: Clean up RendererService.
Much refactoring of the get-property and do-method code.

Also moved a couple of methods out of an inner class. They could go
either place, but all such methods should be in one place.

Also made a few helper methods static since they have nothing to do
with instances.

Renamed a method to clarify that a string arg is used to make a
GVariant of type "object path".