NativeLibs: Use version 3.12.0 of jhbuild
Recall that in our build.sh script, we clone a copy of jhbuild, apply
a patch to it, build it, and then use our custom version of jhbuild
to build all the native libraries.

Unfortunately, the latest version of jhbuild does not work for us;
something in the jhbuild configuration file format has changed, and
it's not clear how to get around that.

With this change, we check out version 3.12.0 of the jhbuild source
before building it.